move user shows routes inside users
