+++ server.js
@@ -1,4 +1,4 @@
-require("dotenv").config({path: "./.env"});
+require("dotenv").config();
 const mongoose = require("mongoose");
 const express = require("express");
 const passport = require("./services/passport");
@@ -24,7 +24,6 @@ app.use(bodyParser.json());
 app.use("/api/auth", require("./routes/auth"));
 app.use("/api/users", require("./routes/users"));
 app.use("/api/shows", require("./routes/shows"));
-app.use("/api/users/shows", passport.authenticate("jwt", {session: false}), require("./routes/user-shows"));
 app.use("/", require("./routes/docs"));
 
 app.listen(parseInt(process.env.PORT), () => console.log(`Listening on port ${process.env.PORT}`));
